gpu: move start_backend from main.rs into lib.rs
This was an oversight - for the user to be able to use the crate as a
library, the start_backend needs to be part of the public API.
